music/requirements.txt
Django==4.2.11
django-htmx==1.17.3
pillow==10.3.0
